Update template_yarte to master
